react: MutationResult.client is always set
As far as I can tell, whenever a MutationResult is provided to the user of
useMutation (or the component or HOC), it always contains a `client`; eg,
OperationData.refreshClient checks this invariant.

MutationResult is also used to type an internal state variable used by
MutationData and useMutation; this one doesn't contain a `client`. This PR
changes that variable to different type without `client`, and the actual exposed
MutationResult now has a non-optional `client`. This makes it match
`QueryResult`.

Now TypeScript consumers of `useMutation` don't need to check to see if `client`
exists before using it.